+++ kaioretry/__init__.py
@@ -12,7 +12,7 @@ from .context import Context
 from .decorator import Retry
 
 
-__version__ = "0.9.0"
+__version__ = "0.9.1"
 
 
 RETRY_PARAMS_DOCSTRING = """